WEB-4454 add edit patient modal
